[level 1] Title: 부족한 금액 계산하기, Time: 0.09 ms, Memory: 71.4 MB -BaekjoonHub

+++ 프로그래머스/1/82612. 부족한 금액 계산하기/부족한 금액 계산하기.java
@@ -7,6 +7,6 @@ class Solution {
         
         if (money - total >= 0) return 0;
         
-        return (long) Math.abs(total - money);
+        return Math.abs(total - money);
     }
 }
